/ ansrole: args: fix missing semicolon in getopt eval line
not quite sure why this was working during testing except when we used
the --help command and it caused an unrecognized option error.  looked
briefly at how this could possiby have worked before the fix, but didn't
get it on first glance, and don't care enough to investigate, it was
obviously a typo and works now for all cases we can throw at it.