- ansrole: play: use a shorter method to generate the vars_files glob
no need to call awk, a list can just be [foo,bar,] in yaml and also we
don't have to remove the vars/ prefix, seems to work fine

this also adds a comment about how we can do it internally with jinja
sometime soon (hopefully) with associated upstream issue number

still don't really get why the vars_files are loaded over and over and
over again but, they are.  must be due to lazy eval but i would think a
contiguous evaluation graph where the predicates didn't change would get
cached.  apparently not so.  but ansible was never really fast or
optimized, was it...