= ansrole: play: remove quotes in defaulted expansion, might be misleading
lest the user think we're inserting a quoted 'true' value to avoid it
being interpreted by the yaml layer as a boolean.  in fact the quote
makes no difference to the expansion -- which will not have quotes,
regardless -- but is something we normally put in shell scripts to make
it clear it's a string, similar to how we do 'echo "foo"' even when
'echo foo' would be the same thing.  just makes it more clear it's not a
variable or keyword.  but in this particular case, given the yaml
context and special treatment of the "magic" strings "true" and "false"
if they are the only string value on the line, we'll just remove these
quotes to make clear we intend for it to be interpreted by yaml as a
bool constant.