+ ignore: add the openssl random seed file which pollutes the homedir
we could use RANDFILE but we'll just leave it where it wants to be for
now.  this file gets deposited when using the ansvarps command,
presumably by elinks.  kind of silly that openssl needs a seed file but
it must make operations faster to keep some state and not have to
regenerate stuff.  but it should be fast enough to just read
/dev/urandom, my system shows 270MB/s from that source