- host: genders: run genders gen only once, gets reused on all targets
no need to have this run as many times as there are hosts, it only needs
to run once.  since it's done only on the localhost via delegate_to, and
the buffer is copied to each target (because no delegate_facts), it only
really needs to run once; the output is the same on all of them and can
be reused.

confirmed with strace that (1) it was executing N copies of
ansible-playbook before the change, and (2) it was executing exaxctly 1
copy after the change.