+ genders: all: move our global dns domain var out to 'all' group
the domain we use has broad application, basically to all our nodes.  if
this changes we'll override it in child groups but it makes sense for it
to belong to all nodes

defining groups requires it either be the top level in genders.yml or
have its own inventory file.  as long as it comes first it will work and
no need to define children since it's an ansible-maintained pseudo-group