+ vars: ipnet: ip_host: do address lookup and use inventory prefix
now that ipnet is not available for all hosts in inventory, as we had
removed it (we don't want to hardcode addresses in inventory, especially
as we are not yet generating host data from inventory but have a fixed
hosts file), we will look up the ip address using gethostbyname().
unfortunately the prefix must be hardcoded, but we have recently added
the 'prefix' variable in inventory because it's the same for all of them
in our setup, we use networks that are all the same size.  not really
satisfied with this solution but for it to be dynamic requires more work
and probably we don't care about it until we are using an ovn setup,
which will not happen for some time