+ host: init: resolv: thwart isc-dhcp-client recreating resolv.conf
during host init, if we found certain directories present, it means the
isc-dhcp-client (dhclient) is installed, and ifupdown uses that
preferentially (it's first in the precedence list, according to
interfaces(5)).  unfortunately that one by default has hooks that set
the system hostname and rewrite resolv.conf which we don't want it to
do.  it sources files, which we place if not already present, and they
will just exit and then aforementioned will never occur.

this is not done if the system is in the noresolv group.  for those we
just want to leave the system resolver configuration untouched.