- all: always use string for octal mode specs, safe and future-proof
there is a long history on this:

  - yaml 1.1 0[0-7]+ -> octal
  - yaml 1.2 0[0-7]+ -> decimal
  - yaml 1.2 0o[0-7]+ -> octal
  - yaml 1.2 not supported in pyyaml for over 15 years but is imminent
  - ansible uses pyyaml which supports only yaml 1.1 so far (202507)
  - ansible in some contexts, like looping/nesting, breaks leading 0
  - ansible in all contexts, if jinja2_native enabled, works leading 0
  - ansible will eventually support yaml 1.2 (authors working on it)
  - ansible always works if a string and will always work with string

due to these peculiarities, and the final point, since strings always
work, we will religiously use strings and always keep a comment that the
fields should be kept in that format.  we will see this bug message.

note some references:

https://github.com/ansible/ansible-lint/issues/ 2965
https://github.com/ansible/ansible/issues/ 78715
https://github.com/yaml/pyyaml/issues/ 116
https://github.com/yaml/pyyaml/pull/ 555
https://github.com/yaml/pyyaml/pull/ 700

ignore the module docs which say it's ok to use leading 0.  for
aforementioned reasons, it is not ok.