+ systune: thpmadvise: defrag: use defer+madvise to avoid slow allocs
observed stalling when flooding the system with requests.  the requests
mostly seem to fail anyways so there must be some kind of timeout.
nonetheless there is no real reason why we need it to be synchronous.
if it's deferred, khugepaged will still merge the pages in the
background eventually, in theory.