+ systune: sysctl: vm: try to thwart memory fragmentation in base knobs
these are pretty good on all systems, we encounter fragmentation issues
all the time historically (and recently!), totally ok to trade a bit of
cpu for that on most systems we have so we will put this in the base
set.  likewise swappiness on linux has always been way too aggressive
out of the box.  25 is a more reasonable value....