+ systune: sysfs: thp: khugepaged: make aggressive against fragmentation
the default settings do not seem to be effective enough on systems where
THP are in use.  in particular, non-trivial allocations end up mostly
failing and since we run with madvise+deferred to avoid stalls, this can
cause an issue where THP are configured, but not used much.  also we
have observed on several systems a scenario where khugepaged just
doesn't run enough to ever make up the difference even in the
background.  so we want it to be more aggressive and have more huge
pages available when they are requested.

since we're already hidden behind madvise() it should be ok to set this
everywhere and not condition on a group.