/ host: init: resolv: dhclient: only override resolv/hostname in hook
the hook actually does important stuff like adding an ip address to the
interface, so we actually have to keep it.  but we can override the two
functions that do the stuff we don't want, in particular they rewrite
resolv.conf and change the system hostname.  we always want to keep our
hostname the same and we use a 127.0.0.x address so that we can.  and we
typicallyuse our own local dns forwarder and configure that separately,
so we never want to change resolv.conf.