/ smemsh: gits: revert delegate_facts back to to false, breaks overrides
the whole point of allowing vars to control the creation of
skel_repos_all was that individual hosts would end up with different
repos, and this would allow overrides to be listed in inventory on hosts
where we needed private branches.  if we delegate facts then there is
only one set of facts that would have to be common to all hosts,
thwarting this mission.