+ ghreltags: add a dmain/driver from devskel to allow invoke/debug from cli
allow to invoke and pass in arguments from the cli rather than going
through ansible.  this will make it much easier to run under a debugger
and test it with various arguments.  we will add a caching layer and
this will be most helpful.  provide the dict on stdin as json, just as
jinja2 would pass it in from the template filter context.