- ghreltags: imports: move all imports, either cli or lib, to preface
if we're invoked from the cli, there are a lot more imports, so we were
doing those extra imports in the dmain section, so it didn't have to
happen when used from ansible.  but really this program takes so long
(ie, does network IO, and overshadowed by ansible itself) that speed
from imports will not matter one bit and just makes it more confusing
when we have to keep track of two different import blocks.  this was
evidenced by commented out sys.foo namespace difference, corrected in
the patch after changing the imports to be at top.