+ ghreltags: add caching of github tags api calls using a python shelf
this is saving on api calls because otherwise it will be redoing all the
repo tags api calls for every host in the play.  for public access the
api limits are quite low; for private we can probably get away with
looking only once per play.  what's nice is that now it doesn't matter
how many hosts there are.

right now this is only caching for the duration of the play but we might
add a refresh= boolean to let it use the last-cached values.

one caveat however: in order for it to work we need to use throttle=1
(although it's implemented in a way that it won't be dangerous if this
is elided, it just will cause tasks to fail if they exceed a timeout
while waiting for a lock), but in practice that shouldn't be too much of
a problem because it's only the set_fact itself (which obtains the api
results) that needs to serialize execution, and all the subsequent ones
after the first will be fully local because they'll hit the cache (the
set_fact is always delegate_to localhost, doesn't make sense to execute
the api calls on remote nodes).  so it should complete quite quickly
even with large number of hosts, when throttle=1.

in theory we could probably up this limit to batch it, as long as a
single batch can always complete by the total sleep time it should work.
so there's room for growth later... we could also implement cache aging
for example, because the cache key is just a time_t value so we can just
add seconds to this to get how long ago the original play was run that
added the entry to the cache.

overall this is quite an enhancement, allows it to scale into the
future, and was fun to implement! very nice...