+ smemsh: skel_repos_excludes: new var to remove repos from default list
sometimes, for example we're working on something and the checkout is
dirty, or, like right now there is a known issue on one of the systems
with one of the repos, who knows when it will be fixed.  most likely we
would add this on the ansrole command line for one-off overrides, but
potentially we would put it in an inventory variable also if it will be
that way for a while or we just don't want that one on the particular
system.  this will allow us to grow the default list too and potentially
put these variables on a group.