+ smemsh: gits: add a skel_repos_submods var for repos to submodule update
certain hosts have certain repositories (like vim) that have a lot of
submodules; on "normal" hosts we want our ~/.vim but without all these
submodules actually checked out.  so the default in
roles/smemsh/tasks/gits.yml is "recursive: false" (kind of a misnomer,
but this does a git submodule update after checkout).  but now we can
set a variable on hosts that should get the full checkout, which will
cause the submodule update to happen.  and leave all the others without
it.

mostly we'll only want this on workstations, so it should be sparse and
we don't need to invent group machinery for it at this time, just add it
as a one-off on certain hosts.