- all: make relevant lookups use query() instead of lookup(wantlist=True)
this is a convenience feature that was added a while ago and we weren't
aware of it.  good idea, was actually just thinking the other day how
silly it was that we had to pass "wantlist=True", but now we don't have
to.

noticed this by accident in a forum post.  it's an ansible update that
happened in the interregnum period after livm, so it probably wasn't in
the high-level release overview, but required reading the detailed
release notes which we did not do during that time.