- accounts: simplify comments and code for boolean condition handling
since we are now using "jinja2_native = True", and since the code in
lib/ansible/plugins/filter/core.py::to_bool() (mapped from 'bool')
passes None unchanged (which evaluates false in a boolean context) and
will pass "" as False, it is safe and makes more sense (since we know
how ansible works now) to remove these comments -- which are stale asof
v2.8.5 and no longer -- and switch to simpler "(val | bool)" for
truthfulness.

note that unlike when it was done, bools passed on the command line with
eg "-e foo=true" are passed as strings (there used to be a cast to bool
for these specific cases).  therefore, we still need the "| bool" filter
to happen, only because those values will be commonly given on the
command line as overrides.  unfortunately, without passing args as json,
there is no way to pass in a boolean or null value.