/ genders: all: add localhost to give it full inventory semantics
when it's implied, we can't use "ansible-playbook --limit=localhost",
and possibly other things like not inheriting configuration variables.
unknown but if we do add it to inventory it becomes a "real" host and
won't get all the "implicit" semantics.

https://docs.ansible.com/ansible/latest/inventory/implicit_localhost.html

we found this out after changing from bin/genders script that used
/etc/genders to dynamically generate inventory, it must have had an
explicit localhost.  once we switched to using yaml inventory, we did
not have an explicit one, so the implicit one took its place and that
fails on --limit=localhost with "limit [...] leaves no hosts to target"
whereas adding it to inventory fixed the problem.

discovered with bin/tfcloudinit, which calls ansrole on localhost, but
even "ansrole localhost:ping" would have failed.