/ terraform: main: omniadmv: ignore when the backing image changes
if we don't ignore, then when the image gets replaced terraform wants to
replaced (ie, we updated it from latest upstream) terraform plans to
destroy all nodes that were based on it.  actually the image isn't
needed after the instance was created; if it's deleted, the thin volume
continues to exist on its own as an orphan and will share blocks with
others that were based on it the original image.