- terraform: localhost: volsz_plex: change default vm size to 32G
even though they're thin volumes, the image upload phase actually makes
a compressed archive of the whole thing and gzips it, which takes
considerably more time when it's larger, even though resulting image
only takes the used space.  so it will be easier when these are smaller.
we can use an external disk if we need to do something unusual on the
node.