/ smemsh: allow to set ssh or pgp keyid false to not do that part
actually the changes we made with truthiness in last few commits worked
against the goal we had originally of allowing those to be set false and
not doing that part, this was the reason for "is
truthy(convert_bool=true)" which lets them be set either to a string
keyid, the string 'false' from command line, or the jinja2 native
boolean value false. but the code would not have worked anyways because
it didn't have conditions on those vars to run the relevant tasks.  this
commit should fix everything to work correctly and skip the right stuff
if set false, and if set to any string it will try to verify and use the
given keyid, and run the ssh role in case of ssh key.