- terraform: locals: rename plexbyhost as gatebyplex, maybe more accurate
this was backwards and 'host' is ambiguous, probably it made sense
though because we call the gateway address the plex address because
that's the address of the node itself.  actually come to think of it
maybe it was named ok because it has other functions than gateway and is
the plex address itself.  but we'll have to think about this and can
easily change it back.