+ typeimgs: new module stores all image variants, master var plexhocs
now we have a master variable, 'plexhocs' that controls the definition
of all our instances in a terraform.tfvars file, and includes os, virt
and type variations in the index string.  to do this we use a module
typeimgs[] that stores all our baked image variations as a second layer
around osimgs[] so we can address them with two levels of indirection
(plexhost and fully qualified image string).

unfortunately this is the only way to do nested for_each addressing in
terraform, as multi-dimension for_each is not possible without
simulating with overloaded string identifiers that get packed with the
multiple dimensions.  for now we'll use true multi-dimension using
modules, which allow one dimension each.

however at the end this allows us to have a single sparse nested object
map that defines all our ad-hoc plexnode instances on all plexhosts,
which is very handy.