- lxdbake: factor out bake params into a yaml dict merge+override
rather than duplicate, better to keep these as an anchor.  we have to
suppress warnings because we use a different 'baketime' so we have to
override it, and ansible warns about the "duplicate" key.  it really
would have no way to know because that's probably in the yaml layer and
ansible never sees the '<<:' key.