- ark: tf: add some unused terraform into an amalgamation .tf file
we like to save these for later in case some idioms are useful.  tired
of having to ignore it in the main terraform directory and not check it
in so we'll just put it in the ark and add to it later if we move things
out.  we can re-examine if we want to even keep these (some of which are
broken and that's why they're unused) later.