- terraform: vars: plexhocs: change to a local so we can use expressions
too limited that we can't even have a range of nodes in a plexhocs
variable, no expressions are allowed and it has to be completely static.
so we'd have to use [1, 2, 3, 4] and not range(1, 4).

and we really don't need to pass it in anywhere or use it outside of the
main module anyways so, really it *is* a local.