/ lxdbake: play: terraform: run on controller only, not remote lxd nodes
this would not work for remote lxd nodes that weren't the controller
itself, terraform is not setup to run anywhere else besides the
controller.  it happened to work when testing terraforming on the
controller because it was only connecting to itself, had setup account,
etc.  but our model is to run terraform incus on the controller only.