- rc: enable strict hostkey checking for ssh connections by default
didn't realize it but hostkey checking was disabled long ago, in fact
it's disabled all the way at the beginning of the repo (2015).  this was
probably before we did hardcoded blocks of address space with precreated
keys, and we had a chicken-egg problem.  using static files and
precomputed keys solved most of these problems.  it creates other ones
of course but this is a good workflow for now and while the keys are in
the repo, they're encrypted and this works for the time being.  and we
have a static known_hosts file so there's no need to disable
strictHostKeyChecking, which has commensurate security issues as well.

we'll use dnssec and sshfp at some future point...