+ bin: tfapply: new wrapper for terraform apply, less noisy
the default output is rather verbose, and warns about things we don't
care about (and warnings cannot be disabled), but -concise flag does not
show the resources.  maybe we have ignored things in this wrapper we
should not ignore, but we'll see how it works in practice.  quite simple
to write, did not take long.