- tf: typeimgs: make fimgs static to persist baked images with no instance
the full-image list has to be static, because if we calculate it from
inventory variables, then images will be destroyed for instances which
would use them but for which there are no dependants defined at the
present time.

wish there was a lifecycle mark like "prevent_destroy" but which would
not generate an error when destruction is called for, it would simply
make destruction a no-op.  there is no such lifecycle flag in terraform
at the moment, unfortunately.