+ ubunode: lxdbake: add logic for incus container bakes, terraform u22c base
the containers don't need boot role run (which deals with grub/kernel
params), certain packages only useful for hardware or device control,
or network time protocol.  other than that the systems bake pretty
normally.  we add support for lxdbake, terraform and ubunode roles to
properly bake lxc container images.  they bake in about 4.5m vs 8m for
vme images, and they provision in about 8s compared to ~35s for vme
images.  this is not really as fast as we would have hoped, it's unclear
why it still takes 8s, given that these are thin snapshots of the base
image and create should be negligible; we need to debug this later.  it
might be doing a second reboot the first time due to cloud-init, just
like the vmes are doing

they also still take about 114M RSS, we can whittle quite a few things
down from this (such as atd), especially if we were willing to use
systemd for everything, it could replace cron, rsyslogd, atd, and we
could get rid of postfix using a smart relay that didn't require a
persistent daemon.  have to think about whether we actually want to do
this.  we made task e5325068 to track this in the meantime.