- jinja2: 219up: rewrites for no more j2 exts, no methods, no mutuations
have to rewrite some idioms to accommodate new limitations in 2.19, see
ansible issue 85202 and 85205.  they took away list.append() and the j2
'do' extension (in fact all extensions), but we can still get some use
back with list = list + item within a namespace.  if we had to rewrite
our iptables rules it would be a lot harder.  that hasn't been merged
into setup yet but it's going to be a daunting task...