+ versions: bump all collections, pypis and ansible-core to v2.19
this is a big one, quite a bit has changed upstream.  recent commits
have fixed most of the issues but there are some more upcoming.  the
error messages are a lot better now, for sure.  but there are some
regressions, especially reduced capability in templates.

we needed to use rc1 because there is a bug in 2.19.0 template lookup
that fails, where template module succeeds.  the fix brings it back to
parity and is vital for us as we use it in a few places.