/ tf: main: typeimgs: refer only to var, not resource, to avoid dependency
if we leave the dependency, then when we destroy the bakenode by tofu
apply -destroy, anything built with it will also be destroyed, and
using -target does not avoid this, nor does lifecycle ignore all.  any
dependent resource is destroyed in destroy mode.  so just referencing
the bakenode's name attribute means any nodes that baked with it will be
destroyed.  this is most unfortunate; there should be a lifecycle arg to
avoid dependent destruction, but there is not.

before implementing this, baking a new image would destroy previously
baked images, which was bogus.