- tf: main: add and use data sources rather than resource references
for projects and profiles, we have only the resources, not any
corresponding data sources.  this would avoid destruction of a node if
its project or profile went away, for example (we would want that
destruction to be explicit).

however we might remove these entirely and just use strings because they
don't really add anything: the string is already present in the resource
reference and is only really used for a name.  so it just adds length
and boilerplate and doesn't remove any duplication.  but we'll do that
in a future commit if we do.