+ tf: imgdata: new module for baked image data sources, avoids dependency
by using data.incus_image rather than resource.incus_image, it avoids
recreate of provisioned nodes on destroy of the typeimg when we rebake,
instead it's only affected via the lifecycle argument.  targeted
destruction of a referenced image will not cascade to the referencing
object as well.

we don't want created nodes to be destroyed if the profile goes away...
in general we do targeted destruction.  this is arguably not how
terraform was designed, but we mostly have snowflake and/or
lab/ephemeral use cases at this point and want destruction to be mostly
explicit.

tried keeping this data source in the timeimgs module, and changing the
timeimgs output to use the data source, but it retains a reference to
the module in instance dependencies if we do this.  unclear why since it
doesn't have to, but moving it to a separate module did work.  we'll
have to just use typeimgs as the bake module that creates the images,
and imgdata as the "extraction" module that provides references.