- ubunode: set_node_facts: add convenience facts about distro and version
we will use these a lot so it helps to have them and make the
conditional expressions shorter.  maybe we want to put this in
vars/main.yml or vars/something.yml eventually because they will be
useful in a lot of roles, but they're still ubuntu-specific and mainly
used here so for the moment we'll keep them in ubunode defaults.