/ cloudinit: do ssh.socket preset disable and ssh.service enable in bootstrap
u24 has race issues during provisioning when we want to remove sshd from
socket activation.  for some reason this does not happen on u22 although
that distro has moved to socket activation also.  unclear why but we
seem to have to do this to get it to work reliably, otherwise it only
works sometimes due to only sometimes winning the race, presumably
between disabling the socket activation and starting the ssh daemon and
for some reason systemd gets confused and "loses track" of which ssh
daemon it should be keeping track of.  it's unclear and we ran out of
patience debugging it so we focused more on getting it to work.