/ ubunode: no_ssh_socket_activation: be more careful about reloads
daemon reload does it *before*, not after, according to docs.
furthermore, we don't want to restart ssh if it's running and reload
seemed to also have issues, there seems to be various race conditions
with disabling socket activation and enabling the service all at once.
after experimentation this was needed to get it to work reliably.