+ bin: tfdeps: new script that displays terraform resource dependencies
the "tofu state show" command does not actually list dependencies,
strangely, and cannot figure out any other way to do it.  but they are
essential for debugging issues with destroy/targeting.  tired of going
and looking in the state file so just made a jq script to do it.

there's probably a way to avoid the extra awk call at the end for
formatting, but could not quickly figure out how so, it's harmless
enough to leave it in there...