^ tf: data: remove unnecessary data sources for project and profile
replace these with simple strings, otherwise it's a whole lot of
boilerplate that actually just includes the string anyways.  terraform
is definitely not known for being DRY.

we do add a local variable for the main project used by all our nodes.
at the moment we are not using different projects besides this one and
the default (where images and other "global" resources like that are).
in the future we may make further use of projects, but not right now.