- tf: vars: move instance disk base size from local to tfvars
this allows us to override, and keeps static configuration where it
belongs.  we also eliminate all the commented out stuff and replace with
a more concise explanation, eliminating the second value because we
won't be using it with lvm for the reasons mentioned in the link (and
for simplicity).  although maybe we just want to set this in the pool
default rather than the profile also, but this way at least it's visible
by inspecting [derived] instance config, and won't depend on the storage
driver if we change to eg zfs later.