+ tfvars: move plexhocs to tfvars and support different formats
comments are self-explanatory but support different methods of
specification.  it's not as flexible as a local var in that they all
have to be specified in the same way, but at least there are several
ways to do it and now it's defined separately from the logic used to
construct the eventual data structured used by terraform.