+ tf: main: plexhocs: cloudinit: no package update stuff on firstboot
the default for package_update is true, but package_upgrade is false.
we don't want to do anything because we maintain these to be relatively
current by frequent re-baking.

and we'll put both keys into the yaml even though one is unnecessary,
just to make explicit how we want to operate.