+ ubunode: facts: add real is_debian, and have uNplus facts include them
we want to (1) be able to make a distinction between "real" debian and
ubuntu proper, and also we want to have version-gates that work on both.
every ubuntu version is based on a specific debian version and they have
essentially the same shared libraries.  so we'll (1) add a real
is_debian (now that we have is_debdist to mean both ubuntu or debian),
and (2) add the right debian version to u2[024]plus variables.