/ gpg: agent: rc: set the log file and home dir with absolute paths
log paths did not seem to expand tildes.  we did a lot of changes at the
same time, unclear if these actually might get expanded but other
changes made it start logging.  nonetheless we're sure it works now and
no harm having these using the expanded form.  this is a ver specific
repo and already has hardcoded name of the setup user and doubtful we'll
have the path somewhere else than /home/setup/, we can fix it at that
time if we do because it will be immediately obvious that gpg-agent will
fail to start and ansvault will not work so secrets will not be
obtainable.