/ ansvault: use gpg via pty so first invoke can be from "sudo su - setup"
when we come via sudo, the default is use_pty, and it means the terminal
will be owned by root.  this means $GPG_TTY will refer to a terminal
that cannot be read.  which means the password won't work without
loopback mode, and even with that mode, cannot be cached.  we could work
around by doing the first invocation via a temporary "script /dev/null"
session to get the password cached, but we keep forgetting this and
spending a lot of time investigating failures.  so instead we make a
wrapper that allocates a pty and always use that for ansvault, this way
we don't ever have to rememeber anymore.

pretty dumb that gpg cannot do this itself but it's been this way for
more than a decade and upstream is always espousing the "security" of
forcing people to do it.

we could also disable use_pty in sudo for root, it is the default
upstream (see sudo-project/sudo issue 258).  but this is better for
other reasons (the security reason cited in the issue is dubious, but
also it's just better to not be keeping the same tty across privilege
boundary).  but this workaround is the winner for now.