/ ansvault: gpg: fix perm warning suppress by correcting pass of argv[0]
zeroeth arg is passed twice.  first time as the path to look up.  second
as the ostensible name of executable.  which can be an executable name
only and doesn't have to be the actual path.  if we don't pass this as
the first argument ("zeroeth") then the actual first argument gets
consumed, so our --no-permission-warning never takes.