/ ansvault: gpg: wrapper must elide pty also when stdin double-wrapped
if we already ran one of our scripts that reopens tty when stdin is
available (such as tfcloudinit does, getting json on stdin, then it will
end up calling ansvault from within, which calls bin/gpg), then for some
reason (which needs further research), the double-wrapped script's
sys.stdin is None.  so that generates a crash, as it is unexpected.

for the moment we work around this by detecting this condition and using
the "real" gpg when that happens.  just like we were doing it already if
stdin was available by select(), we also do it now if stdin is
unavailable by virtue of being None.  but really we have to sort out (1)
why stdin is doing this and (2) get stdin to work anyways using a pty,
if it's available.

however those issues are not interfering with operations so for the time
being, so we will just skip our wrapper when either of these conditions
happen, for now.