/ vars: ansible_managed: do not include the comment character
allow it to be used in other contexts and fix some contexts where it was
misused

we cannot use it in static file contexts because there is no templating
so we do not have variable interpolation available.  probably these
contexts should be switched to templates at some point but files are
quite simple so we'll just have to duplicate the "ansible-managed"
string within these files by hand for the moment.  since we've been
using the same idiom for years now this probably won't be a problem.