/ ssh: disable_non_rsa_hostkeys: make explicit to avoid upgrade script keygen
just removing the keys is not enough, the upgrade script will generate
missing ones *unless* the HostKey is specified and then.  don't really
like how we did this because you have to "know" the list of them and
remove the ones except the one we want.  nonetheless this will work for
at least several years, most likely.