+ systune: brouter: add the concept of a routing bridge for kubernetes
k8s needs the L2 packets to be passed into iptables for further
processing, and the kernel module needs to be loaded explicitly...
nothing will compel it.

we allow this to be specified as a role argument (which is used in the
kubernetes node type play), and also as a group, so it will work in
different contexts, and we don't have to add all hocnodes to some kind
of group to get it to work, but rather can just do it explicitly at
provision-time, while other nodes can use the group if they wish.