^ ubunode: pkg: emacs: we will kill off emacs-nox as it's unsupported
the emacs-nox package is only in the universe pool and therefore does
not get security updates.  it was only added for old $BOSS to have his
editor omnipresent, but we ourselves never use emacs, and it's a beast.
leaving it commented out for now in case it comes up later, which is
doubtful, but vim users are hard to find so maybe some day.