- tf: lifecycle: ignore_changes: add user-data, is for provision-time only
once the node is provisioned, typically the user-data is not used
anymore except on the first deploy boot.

there is possibility to use it afterwards, by bootcmd, since
cloud-init runs that every boot, but we only use runcmd

that's not really our model, to rely on inter-boot changes.  for kube
nodes we are pure ephemeral nature, they don't really have intent to
have multiple boots and do anything different without a reprovision.  we
might implement this later for apiserver nodes, which are static "seeds"
and can potentially bootstrap entirely different clusters, we might do
that later.  but nothing that we're looking to do now, and even in that
case, pure reprovision is probably better anyways.