- postfix: tmpl: gmek: work with different default interfaces than eth0
my laptop was on a wlan0 when out, which made the ansible stop working.
this will use the primary interface of hardcoding eth0.  however what
this means is that we'll have to keep writing the file every time we
change our default interface, which shouldn't be that often, but seems
like we should figure out a better way.

looked into this somewhat and there really isn't a better way though,
besides binding on all addresses.  the listen address unfortunately if
it does not include the smtp bind address then it just listens on all
interfaces.

so smtp bind address has to be specified.  but at least in our case on
plex nodes it always routes via meklar so we just give a lookup of the
hosts's terra network IP using our 'addrof' python filter which does a
trivial host entry lookup (strangely unavailable in ansible out of the
box...)