/ tf: plexvars: plexhocnodes: normalize each test arg so try() works
try() expects all its args to be the same type.  since one of them is a
list, it means they all have to be, so even as strings we'll need to
pass them as one-member lists.  seem to recall testing all three
variations after writing the "big try" but the evidence would seem to
indicate that we did not.