+ tf: plexhocs: wait_for: poll ssh port with ansible before proceeding
we have a local-exec provisioner that calls ssh, it does not use
remote-exec and configure ssh herein but relies on external ssh
configuration to be working already.  so we run a provisioner which
requires that ssh is up on the new incus_instance.

we have found with a high number of nodes starting at once, the system
can get network up (which is the terraform wait_for condition)
significantly before ssh starts listening, so there's a race and the
provisioner will usually win at high loads.  so it gets a connection
refused and then thinks the provisioner didn't work, but really it just
had to wait longer.