+ tf: kube: add kubernetes support, 1st plex node is master, rest are slaves
we have two plexes for our lab setup, now they are each kubernetes
kubes.  when fooplexN are provisioned, as long as they are image type
"kube", they will be either a maser (fooplex1) or a slave (fooplex2+).
everything will provision and setup automatic, done through cloud-init,
depending on the system role kubeadm init (master) or kubeadm join
(slave).

have not yet figured out how to make fooplex2+ depend on fooplex1.
unfortunately it's deemed as self-referential or cyclic depending on
whether we use a terraform_data indirection to try to do it, the result
is always the same.  without making a separate resource completely just
for the master node (and then making the plexhocs be 2-256 rather than
1-256), there seems truly no way to do it.  however as the operator we
can "know" this for now that if we regenerate a new master then
necessarily the slaves will be defunct and require replacement to rejoin
(or a mass re-join operation).  likewise we have to "know" that the
master must exist and be fully initialized prior to any of the slaves
provisioning.  meanwhile we'll consider how to refactor the plexhocs
resources to allow most of their configuration to be specified outside
the resource block and referred to within the resource block, to reduce
duplication.