- postfix: main: allow bind on all addresses, deal with it in firewall
we will have to deal with this in the firewall.  the problem is there is
a race condition with the socket bind: postfix is supposed to bind on a
certain interface, which it could do at start, but then when it uses
smtp to relay it has to bind to the tunnel address and the the tunnel is
not up yet.  howver, if we don't specify the smtp_bind_address then
smtpd listens on all interfaces even if inet_interfaces only gives one,
not quite clear why this happens but it has something to do with not
knowing the right interface to go out.  if you tell it with
smtp_bind_address, then that interface must exist, which it does not
until openvpn starts.  even if we put a systemd override and make
postfix@-.service After=openvpn@ourconfig then it still doesn't work
because the interface apparently is not actually operational yet and
also this relies on specific details of the node that we really don't
want postfix to know anything about.

ideally postfix would just use IP_FREEBIND socket option so the
interface would not have to exist at the start, but it does not and that
string is not present anywhere in the source code.  it's an old software
and used mostly in commercial settings with dedicated machines, so it's
unlikely this will be added.  that ship has already sailed...

instead we will let it bind on INADDR_ANY and use firewall restrictions
to control what can connect on port 25.  it's only the plexnodes that
will be connecting there.  this gets rid of race condition at start
where tunnel does not exist yet so the bind doesn't work.  even if we
layer it and say postfix needs to